Split ChatPage to separate components
