+++ frontend/src/app/chat/[id]/chat-page.tsx
@@ -1,18 +1,14 @@
 "use client"
 
-import { FC, useEffect, useRef, useState } from "react"
-import ReactTextAreaAutosize from "react-textarea-autosize"
+import { FC, useEffect } from "react"
 import { useParams, useRouter } from "next/navigation"
-import dayjs from "dayjs"
 import { useDispatch } from "react-redux"
 
-import Markdown from "@/components/markdown/markdown"
-import { useChatMessages, useMessage, useSendMessage } from "@/api/chats/messages"
-import { useIsUserOnline, useUser } from "@/api/users/hooks"
-import { useChat } from "@/api/chats/chats"
 import { chatActions } from "@/store/chats"
-import { useMe } from "@/api/auth/hooks"
-import UserLink from "@/components/UserLink"
+import { useChat } from "@/api/chats/chats"
+import ChatInfo from "@/components/chats/ChatInfo"
+import ChatMessages from "@/components/chats/Messages"
+import WriteMessageForm from "@/components/chats/WriteMessageForm"
 
 const ChatPage = () => {
   const chatId = Number(useParams().id)
@@ -58,162 +54,6 @@ const Chat: FC<{ id: number }> = ({ id }) => {
       <ChatInfo userId={chat.companionId} />
       <ChatMessages chatId={chat.id} />
       <WriteMessageForm chatId={chat.id} />
-    </div>
-  )
-}
-
-const WriteMessageForm: FC<{ chatId: number }> = ({ chatId }) => {
-  const [input, setInput] = useState("")
-
-  const formRef = useRef<HTMLFormElement>(null)
-  const inputRef = useRef<HTMLTextAreaElement>(null)
-
-  const sendMessage = useSendMessage()
-
-  useEffect(() => {
-    inputRef.current?.focus()
-  }, [chatId])
-
-  return (
-    <form
-      ref={formRef}
-      className={"mb-6"}
-      onSubmit={async (e) => {
-        e.preventDefault()
-        sendMessage(chatId, input)
-        // await mutateMessages()
-        setInput("")
-      }}
-    >
-      <ReactTextAreaAutosize
-        ref={inputRef}
-        className={"textarea w-full my-3"}
-        onChange={(e) => setInput(e.currentTarget.value)}
-        onBlur={() => setInput(input.trim())}
-        value={input}
-        onKeyDown={(e) => {
-          if (e.key === "Enter" && !e.shiftKey) {
-            e.preventDefault()
-            const trimmedInput = input.trim()
-            setInput(trimmedInput)
-            if (trimmedInput === "") {
-              return
-            }
-            formRef.current?.dispatchEvent(new Event("submit", { cancelable: true, bubbles: true }))
-          }
-        }}
-        required
-      />
-      <button className={"button disabled:opacity-50"} type={"submit"}>
-        Send
-      </button>
-    </form>
-  )
-}
-
-const ChatInfo: FC<{ userId: number }> = ({ userId }) => {
-  const { user } = useUser(userId)
-  const isOnline = useIsUserOnline(userId)
-
-  if (!user) {
-    return <div>loading...</div>
-  }
-  return (
-    <h1 className={"text-2xl mb-auto pb-2 border-b"}>
-      Chat with{" "}
-      <UserLink userId={user.id}>
-        <span className={"font-bold clickable"}>{user?.username}</span> {isOnline ? "🟢" : "🔴"}
-      </UserLink>
-    </h1>
-  )
-}
-
-const ChatMessages: FC<{ chatId: number }> = ({ chatId }) => {
-  const { chatMessages } = useChatMessages(chatId)
-
-  const scrollRef = useRef<HTMLDivElement>(null)
-
-  const isOnBottom = useRef(true)
-
-  // TODO: add lazy loading
-
-  // TODO: add dates between days
-
-  useEffect(() => {
-    // TODO: improve this
-    setTimeout(() => {
-      scrollRef.current?.scrollTo(0, 0)
-    }, 100)
-  }, [chatMessages])
-
-  if (!chatMessages) {
-    return <div>loading...</div>
-  }
-
-  return (
-    <div
-      ref={scrollRef}
-      onScroll={(e) => {
-        // scroll is reversed, so scrollTop is always negative
-        isOnBottom.current = -50 < e.currentTarget.scrollTop
-      }}
-      className={"overflow-y-auto flex flex-col-reverse gap-1.5 items-center"}
-    >
-      {chatMessages
-        .slice()
-        .reverse()
-        .map((messageId) => (
-          <MessageComponent key={messageId} id={messageId} />
-        ))}
-    </div>
-  )
-}
-
-const MessageComponent: FC<{ id: number }> = ({ id }) => {
-  const { user } = useMe()
-  const { message } = useMessage(id)
-
-  const ref = useRef<HTMLDivElement>(null)
-
-  if (!message) {
-    return null
-  }
-
-  if (message.authorId === -1) {
-    return (
-      <span
-        className={"mx-auto italic"}
-        title={dayjs(message.timestamp).format("YYYY-MM-DD HH:mm:ss")}
-      >
-        {message.content} at {dayjs(message.timestamp).format("HH:mm")}
-      </span>
-    )
-  }
-
-  return (
-    <div
-      ref={ref}
-      className={
-        "py-1 px-3 rounded w-fit flex flex-wrap max-w-[85%]" +
-        " " +
-        (message.authorId === -1
-          ? "mx-auto"
-          : user?.id !== message.authorId
-          ? "rounded-bl-none dark:bg-gray-800 bg-blue-200 mr-auto justify-end"
-          : "rounded-br-none dark:bg-gray-900 bg-gray-200 ml-auto justify-end")
-      }
-    >
-      <Markdown
-        className={"prose-img:max-h-[50vh]"}
-        content={message.content}
-        fallback={message.content}
-      />
-      <span
-        className={"ml-2 mt-auto text-sm opacity-75"}
-        title={dayjs(message.timestamp).format("YYYY-MM-DD HH:mm:ss")}
-      >
-        {dayjs(message.timestamp).format("HH:mm")}
-      </span>
     </div>
   )
 }
